requirements.txt
IPython
uvicorn
fastapi
numpy
h5py
requests
pyserial
pyyaml